+++ misc/metrics/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.12.0 [unreleased]
+# 0.12.0
 
 - Update to `prometheus-client` `v0.19.0`. See [PR 3207].